feat(env-check): changest, tests, refactor
