--- .changeset/brave-moons-search.md
@@ -1,5 +1,0 @@
----
-'@sap-ux/environment-check': minor
----
-
-New module @sap-ux/environment-check
