+++ .changeset/great-ghosts-yawn.md
@@ -1,0 +1,6 @@
+---
+'@sap-ux/axios-extension': minor
+'@sap-ux/environment-check': minor
+---
+
+New module @sap-ux/environment-check. Small bug fix in @sap-ux/axios-extension
