+++ packages/environment-check/test/checks/utils.test.ts
@@ -1,8 +1,8 @@
-import { findAllPackageJsonFolders } from '../src/utils';
+import { findAllPackageJsonFolders } from '../../src/checks/project-utils';
 import { join } from 'path';
 
 describe('Test findAllPackageJsonFolders:', () => {
-    const sampleWorkspace = join(__dirname, 'sample-workspace');
+    const sampleWorkspace = join(__dirname, '..', 'sample-workspace');
     // Mock setup
     const wsFolders = [sampleWorkspace];
     test('Find all projects with package.json', async () => {
